# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API za praćenje kompetencija",
        "description": "Dokumentacija obuhvata rute za upravljanje kompetencijama, verifikacijama i autentifikaciju korisnika.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/competency-types": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Lista svih tipova kompetencija",
                "operationId": "9cc3e09eebc6051536b0ca7f9dd0a651",
                "responses": {
                    "200": {
                        "description": "Uspešno dobavljena lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Dodaj novi tip kompetencije",
                "operationId": "130bff303780e42c35e7a1f37ad389ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Programiranje"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešno kreirano"
                    },
                    "422": {
                        "description": "Greška u validaciji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/competency-types/{id}": {
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Ažuriraj tip kompetencije",
                "operationId": "775c83305dc6adc7491ca26f38f8a295",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dizajn"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirano"
                    },
                    "404": {
                        "description": "Tip nije pronađen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Obriši tip kompetencije",
                "operationId": "5c72dca40d302b8a0b2800534e70f21d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno obrisano"
                    },
                    "404": {
                        "description": "Tip nije pronađen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/institutions": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Lista svih institucija",
                "operationId": "32ef01ce0a6e2578738f42463f66c478",
                "responses": {
                    "200": {
                        "description": "Uspešno dobavljena lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Dodaj novu instituciju",
                "operationId": "e28467aab770e2d18c848aa01045d4a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Univerzitet u Beogradu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Institucija uspešno kreirana"
                    },
                    "422": {
                        "description": "Greška u validaciji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/institutions/{id}": {
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Ažuriraj podatke o instituciji",
                "operationId": "98e02a95fe073abbbe97909a8051ce25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novi Sad IT Akademija"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirano"
                    },
                    "404": {
                        "description": "Institucija nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Obriši instituciju",
                "operationId": "cd99565e970d793f5a5991ac151fbaa4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Uspešno obrisano"
                    },
                    "404": {
                        "description": "Institucija nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/system-logs": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Pregled svih sistemskih logova (Samo Admin)",
                "operationId": "08a1343330dcd6b9c843a5d6d7ac82c0",
                "responses": {
                    "200": {
                        "description": "Lista svih akcija u sistemu sa podacima o korisnicima"
                    },
                    "403": {
                        "description": "Zabranjen pristup - niste admin"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Ručno kreiranje log zapisa",
                "operationId": "2e5ed08f9ad79034c809b82a8ea72f38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "example": "Manual Audit"
                                    },
                                    "entity": {
                                        "type": "string",
                                        "example": "User Management"
                                    },
                                    "entity_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Provera statusa korisnika"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Log uspešno kreiran"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}